git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@1860 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/registrar-global.c
@@ -489,7 +489,10 @@ static void statisticsCallback(struct Dispatcher* dispatcher,
             userTime / 1000000.0,
             systemTime / 1000000.0,
 
-            peers, pools, poolElements, ownedPoolElements,
+            (unsigned long long)peers,
+            (unsigned long long)pools,
+            (unsigned long long)poolElements,
+            (unsigned long long)ownedPoolElements,
 
             registrar->RegistrationCount,
             registrar->ReregistrationCount,
